+++ regression/python/while/main.py
@@ -3,7 +3,7 @@ count : int = 0
 while count < 5:
     count = count + 1
 
-div: int = 1/count
+div:float = 1/count
 
 def foo() -> None:
     x = 2
